Refactor/album update

+++ reviewSite/album/cron.py
@@ -6,8 +6,9 @@ import pandas as pd
 from konlpy.tag import Okt # 형태소 분석기 Okt
 from collections import Counter
 import os
-import sqlite3
 from datetime import datetime
+from album.models import Album
+import shutil
 
 
 def update_info():
@@ -63,177 +64,18 @@ def update_info():
     reviews_df['tokenized'] = reviews_df['review'].apply(tokenizer.nouns)
 
 
-    # DB 연결
-    conn = sqlite3.connect(DB_PATH)
-    cur = conn.cursor()
-
-
     # 영화 랭킹 순으로 긍정, 부정 워드 클라우드 만들기
+    is_positive_comment_sufficent = False
+    is_negative_comment_sufficent = False
+
     for i in range(0,len(movie_ranking_df)):
 
         matched_title_reviews_df = reviews_df.loc[reviews_df['movie_title']==movie_ranking_df['title'][i]]
 
-        # 해당 영화에 리뷰가 없을경우 continue
-        if len(matched_title_reviews_df) == 0:
-            continue
+        # 긍정 리뷰가 1개 이상이라면 긍정 워드클라우드 생성
+        if len(matched_title_reviews_df[matched_title_reviews_df['label']==1]['tokenized'])!=0:
+            is_positive_comment_sufficent = True
 
-        elif len(matched_title_reviews_df[matched_title_reviews_df['label']==1]['tokenized'])!=0 and len(matched_title_reviews_df[matched_title_reviews_df['label']==0]['tokenized'])!=0: 
-            positive_reviews = np.hstack(matched_title_reviews_df[matched_title_reviews_df['label']==1]['tokenized'].values)
-            negative_reviews = np.hstack(matched_title_reviews_df[matched_title_reviews_df['label']==0]['tokenized'].values)
-
-            positive_count=Counter(positive_reviews)
-            negative_count=Counter(negative_reviews)
-
-            rank_text_p=dict(positive_count)
-            rank_text_n=dict(negative_count)
-
-
-            # 불용어 제거
-            temp_dic_p={}
-
-            for key, value in rank_text_p.items():
-                if key not in k_stopword:
-                    temp_dic_p[key]=value
-
-            temp_dic_n={}
-            for key, value in rank_text_n.items():
-                if key not in k_stopword:
-                    temp_dic_n[key]=value
-
-
-            wordcloud_p = WordCloud(
-                font_path = WORD_CLOUD_DATA_PATH + 'NanumGothic.ttf',
-                width = 500,
-                height = 500,
-                background_color = "white",
-                mask = mask_p
-            )
-
-            wordcloud_n = WordCloud(
-                font_path = WORD_CLOUD_DATA_PATH + 'NanumGothic.ttf',
-                width = 500,
-                height = 500,
-                background_color = "white",
-                mask = mask_n
-            )
-
-            wordcloud_p = wordcloud_p.generate_from_frequencies(temp_dic_p)
-            wordcloud_n = wordcloud_n.generate_from_frequencies(temp_dic_n)
-            image_colors_p = ImageColorGenerator(mask_p)
-            image_colors_n = ImageColorGenerator(mask_n)
-
-
-            # Title 업데이트
-            title = movie_ranking_df['title'][i]
-            query = "UPDATE album_album SET title = '%s' WHERE id = '%s'" % (title, str(i+1))
-            cur.execute(query) # 제목 업데이트하는 쿼리 실행
-
-            query2 = "UPDATE album_album SET dt_created = datetime('now') WHERE id = '%s'" % (str(i+1))
-            cur.execute(query2) # 업데이트 날짜로 바꾸는 쿼리 실행
-
-            conn.commit() # DB에 반영
-
-            #긍정리뷰 이미지 저장
-            plt.figure(figsize=(10, 10))
-            plt.imshow(wordcloud_p.recolor(color_func=image_colors_p), interpolation="bilinear")
-            plt.axis("off")
-
-            total_good_img_save_path = IMG_SAVE_PATH + str(i+1) + '.png'
-
-            # 이미지 덮어쓰기가 plt에서는 지원이 안되므로 삭제했다가 저장 새로하기
-            if os.path.isfile(total_good_img_save_path):
-                os.remove(total_good_img_save_path)
-
-            plt.savefig(total_good_img_save_path)
-
-            #부정리뷰 이미지 저장
-            plt.figure(figsize=(10,10))
-            plt.imshow(wordcloud_n.recolor(color_func=image_colors_n), interpolation="bilinear")
-            plt.axis("off")
-
-            total_bad_img_save_path = IMG_SAVE_PATH + str(i+1) + '-1' + '.png'
-                
-            # 이미지 덮어쓰기가 plt에서는 지원이 안되므로 삭제했다가 저장 새로하기
-            if os.path.isfile(total_bad_img_save_path):
-                os.remove(total_bad_img_save_path)
-                
-            plt.savefig(total_bad_img_save_path)
-
-            # 포스터 변경
-            poster_url = movie_ranking_df['img'][i]
-            os.system("curl " + poster_url + " > " + IMG_SAVE_PATH + "%s_poster.jpg" % (str(i+1)))
-
-        # 해당 영화가 부정적인 리뷰만 있는 경우
-        elif len(matched_title_reviews_df[matched_title_reviews_df['label']==1]['tokenized'])==0:
-
-            negative_reviews = np.hstack(matched_title_reviews_df[matched_title_reviews_df['label']==0]['tokenized'].values)
-            negative_count=Counter(negative_reviews)
-
-            rank_text_n=dict(negative_count)
-
-
-            # 불용어 제거
-            temp_dic_n={}
-            for key, value in rank_text_n.items():
-                if key not in k_stopword:
-                    temp_dic_n[key]=value
-
-            wordcloud_n = WordCloud(
-                font_path = WORD_CLOUD_DATA_PATH + 'NanumGothic.ttf',
-                width = 500,
-                height = 500,
-                background_color = "white",
-                mask = mask_n
-            )
-
-            wordcloud_n = wordcloud_n.generate_from_frequencies(temp_dic_n)
-            
-            image_colors_n = ImageColorGenerator(mask_n)
-
-
-            # Title 업데이트
-            title = movie_ranking_df['title'][i]
-            query = "UPDATE album_album SET title = '%s' WHERE id = '%s'" % (title, str(i+1))
-            cur.execute(query) # 제목 업데이트하는 쿼리 실행
-
-            query2 = "UPDATE album_album SET dt_created = datetime('now') WHERE id = '%s'" % (str(i+1))
-            cur.execute(query2) # 업데이트 날짜로 바꾸는 쿼리 실행
-
-            conn.commit() # DB에 반영
-
-            #긍정리뷰 이미지 저장
-            plt.figure(figsize=(10, 10))
-            plt.imshow(img_w, interpolation="bilinear")
-            plt.axis("off")
-
-            total_good_img_save_path = IMG_SAVE_PATH + str(i+1) + '.png'
-
-            # 이미지 덮어쓰기가 plt에서는 지원이 안되므로 삭제했다가 저장 새로하기
-            if os.path.isfile(total_good_img_save_path):
-                os.remove(total_good_img_save_path)
-
-            plt.savefig(total_good_img_save_path)
-
-            #부정리뷰 이미지 저장
-            plt.figure(figsize=(10,10))
-            plt.imshow(wordcloud_n.recolor(color_func=image_colors_n), interpolation="bilinear")
-            plt.axis("off")
-
-            total_bad_img_save_path = IMG_SAVE_PATH + str(i+1) + '-1' + '.png'
-                
-            # 이미지 덮어쓰기가 plt에서는 지원이 안되므로 삭제했다가 저장 새로하기
-            if os.path.isfile(total_bad_img_save_path):
-                os.remove(total_bad_img_save_path)
-                
-            plt.savefig(total_bad_img_save_path)
-
-            # 포스터 변경
-            poster_url = movie_ranking_df['img'][i]
-            os.system("curl " + poster_url + " > " + IMG_SAVE_PATH + "%s_poster.jpg" % (str(i+1)))
-
-        # 해당 영화가 긍정적인 리뷰만 있는 경우
-        elif len(matched_title_reviews_df[matched_title_reviews_df['label']==0]['tokenized'])==0:
-            
             positive_reviews = np.hstack(matched_title_reviews_df[matched_title_reviews_df['label']==1]['tokenized'].values)
             positive_count=Counter(positive_reviews)
 
@@ -259,49 +101,101 @@ def update_info():
             image_colors_p = ImageColorGenerator(mask_p)
 
 
-            # Title 업데이트
-            title = movie_ranking_df['title'][i]
-            query = "UPDATE album_album SET title = '%s' WHERE id = '%s'" % (title, str(i+1))
-            cur.execute(query) # 제목 업데이트하는 쿼리 실행
+        # 부정 리뷰가 1개 이상이라면 부정 워드클라우드 생성
+        if len(matched_title_reviews_df[matched_title_reviews_df['label']==0]['tokenized'])!=0:
+            is_negative_comment_sufficent = True
 
-            query2 = "UPDATE album_album SET dt_created = datetime('now') WHERE id = '%s'" % (str(i+1))
-            cur.execute(query2) # 업데이트 날짜로 바꾸는 쿼리 실행
+            negative_reviews = np.hstack(matched_title_reviews_df[matched_title_reviews_df['label']==0]['tokenized'].values)
+            negative_count=Counter(negative_reviews)
 
-            conn.commit() # DB에 반영
+            rank_text_n=dict(negative_count)
 
-            #긍정리뷰 이미지 저장
+
+            # 불용어 제거
+            temp_dic_n={}
+            for key, value in rank_text_n.items():
+                if key not in k_stopword:
+                    temp_dic_n[key]=value
+
+            wordcloud_n = WordCloud(
+                font_path = WORD_CLOUD_DATA_PATH + 'NanumGothic.ttf',
+                width = 500,
+                height = 500,
+                background_color = "white",
+                mask = mask_n
+            )
+
+            wordcloud_n = wordcloud_n.generate_from_frequencies(temp_dic_n)
+            image_colors_n = ImageColorGenerator(mask_n)
+
+
+
+        total_good_img_save_path = IMG_SAVE_PATH + str(i+1) + '.png'
+
+        # 긍정 워드클라우드가 생성된 경우 이미지 저장
+        if is_positive_comment_sufficent:
             plt.figure(figsize=(10, 10))
             plt.imshow(wordcloud_p.recolor(color_func=image_colors_p), interpolation="bilinear")
             plt.axis("off")
 
-            total_good_img_save_path = IMG_SAVE_PATH + str(i+1) + '.png'
-
             # 이미지 덮어쓰기가 plt에서는 지원이 안되므로 삭제했다가 저장 새로하기
             if os.path.isfile(total_good_img_save_path):
                 os.remove(total_good_img_save_path)
-
             plt.savefig(total_good_img_save_path)
+        # 아닐 경우 흰 이미지로 대체
+        else:
+            shutil.copy2(img_white, total_good_img_save_path) # shutil.copy2(src, dst) : src의 파일 정보를 dst에 복사
 
-            #부정리뷰 이미지 저장
+
+        total_bad_img_save_path = IMG_SAVE_PATH + str(i+1) + '-1' + '.png'
+
+        # 부정 워드클라우드가 생성된 경우 이미지 저장
+        if is_negative_comment_sufficent:
             plt.figure(figsize=(10,10))
-            plt.imshow(img_w,interpolation='bilinear')
+            plt.imshow(wordcloud_n.recolor(color_func=image_colors_n), interpolation="bilinear")
             plt.axis("off")
 
-            total_bad_img_save_path = IMG_SAVE_PATH + str(i+1) + '-1' + '.png'
-                
             # 이미지 덮어쓰기가 plt에서는 지원이 안되므로 삭제했다가 저장 새로하기
             if os.path.isfile(total_bad_img_save_path):
                 os.remove(total_bad_img_save_path)
-                
             plt.savefig(total_bad_img_save_path)
+        # 아닐 경우 흰 이미지로 대체
+        else:
+            shutil.copy2(img_white, total_bad_img_save_path) # shutil.copy2(src, dst) : src의 파일 정보를 dst에 복사
 
-            # 포스터 변경
-            poster_url = movie_ranking_df['img'][i]
-            os.system("curl " + poster_url + " > " + IMG_SAVE_PATH + "%s_poster.jpg" % (str(i+1)))
 
-    # for문 끝나고 DB 연결 해제
-    cur.close()
-    conn.close()
+        # flag 초기화
+        is_positive_comment_sufficent = False
+        is_negative_comment_sufficent = False
+
+
+        # 포스터 변경
+        poster_url = movie_ranking_df['img'][i]
+        os.system("curl " + poster_url + " > " + IMG_SAVE_PATH + "%s_poster.jpg" % (str(i+1)))
+
+
+
+        # 정보를 변경할 앨범 객체 가져오기
+        album = Album.objects.get(id=i+1)
+
+        # Title 업데이트
+        title = movie_ranking_df['title'][i]
+        album.title = title
+
+        # 업데이트 날짜로 바꾸기
+        album.dt_updated = datetime.now()
+
+        # 추천한 유저 리스트 초기화
+        album.positive_voters.set([])
+        album.negative_voters.set([])
+
+        # 추천수 초기화
+        album.positive_votes_n = 0
+        album.negative_votes_n = 0
+
+        # 변경 사항 저장
+        album.save()
+
 
     now = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
     print(f'{now} : Done web page update.') # log 확인을 위해서 종료 메시지 추가
@@ -310,11 +204,11 @@ def update_info():
 def save_csv_from_git():
     now = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
     BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-    
+
     movie_df = pd.read_csv('https://raw.githubusercontent.com/Devcourse-1th-team1-1/back-end-repo/crawling/crawlingFolder/backup/movie_rankings.csv')
-    
+
     review_df = pd.read_csv('https://raw.githubusercontent.com/Devcourse-1th-team1-1/back-end-repo/crawling/crawlingFolder/backup/movie_reviews_total.csv')
-    
+
     movie_df.to_csv(os.path.join(BASE_DIR, 'data/movie_rankings.csv'), index=False)
     review_df.to_csv(os.path.join(BASE_DIR, 'data/movie_reviews_total.csv'), index=False)
     print(f'{now} : save_csv_from_git')
